uFo doors went silent... bork!
 - fix several compiler warnings (casts)
 - BattlescapeGame,
	- change var names
	- went over checkForCasualties, esp. morale modifiers
 - insert/remove debug_Info
 - surface adj. Stores, SaveGameState, SaveState
 - change colors of "saveGame..."
 - add tuSnap counter in updateSoldierInfo()
	- rewrite to fix CTD when selectedUnit != exists
 - fix BattlescapeState::mapClick()
 - getActiveHand() now returns empty string if apropo
 - reinsert my hiddenMovement reveal
 - make check for validUnit in TileEngine::hit()
	- use aLien morale modifier
	- change varnames
 - calculateParabola() change varnames
	- increment attempts by 0.5 instead of 1.0
 - validateThrow() close to calculateParabola()
 - TileEngine::distance() now uses 3dimensions
	- removed the +0.5 for whatever reason.
- formatting etc.

+++ src/Basescape/CraftSoldiersState.cpp
@@ -204,7 +204,7 @@ void CraftSoldiersState::btnUnloadClick(Action*)
 
 	// iterate over all listRows and change their stringText and lineColor
 	Uint8 color;
-	for (int r = 0; r < _base->getSoldiers()->size(); ++r)
+	for (int r = 0; (unsigned)r < _base->getSoldiers()->size(); ++r)
 	{
 		std::wstring craft2 = _lstSoldiers->getCellText(r, 2);
 		if (craft2 == craft1) // if row pertains to this craft
